Move product variant code to its own file
